feat(pom-cli): skipSystemFonts 対応・worker プロセス廃止・publish 設定追加
- pptx-glimpse を 0.11.0 に更新し skipSystemFonts: true を使用
  - システムフォントのスキャンが不要になったため OOM を回避するための
    child process worker 架構を削除し preview.ts に SVG 生成をインライン化
- fonts/ シンボリックリンクを実ファイルに置き換え npm publish で配布可能に
- package.json に prepublishOnly / publishConfig / files (fonts/) を追加
- release.yml に pom-cli の build / fmt:check / lint / typecheck を追加

close #711
