+++ .changeset/pom-cli-publish-ready.md
@@ -1,0 +1,5 @@
+---
+"pom-cli": minor
+---
+
+skipSystemFonts を使い SVG 生成の worker プロセスを廃止
