--- packages/pom-cli/fonts
@@ -1,1 +1,0 @@
-../pom-vscode/fonts
